move atlas.rs into a mod
move Image and related types to own file
